feat: KYC component implementation in claimflow
